WoundSystem cleanup and minor refactors (#6333)
* pass one

* Holy slop this slop is so slop i unslopped some of it (kill linq, make more generic methods, split files) but its still so slop help me

* fix damage spreading between limbs + total linq death + other bugfixes and cleanup

* prediction moment

* bro

* Nuke damage redirection, cant get it to work quite right, ill do it in another PR

* yaml fix

* Fuck nukeops

* Bucket ig

* other way round fuck off

* wow

---------
