+++ Content.Shared/_Shitmed/Surgery/Wounds/Components/WoundableComponent.cs
@@ -14,6 +14,7 @@ using Robust.Shared.Audio;
 using Robust.Shared.Containers;
 using Robust.Shared.GameStates;
 using Robust.Shared.Prototypes;
+using Content.Shared._Shitmed.Medical.Surgery.Wounds.Systems;
 using Robust.Shared.Serialization;
 
 namespace Content.Shared._Shitmed.Medical.Surgery.Wounds.Components;
@@ -73,10 +74,17 @@ public sealed partial class WoundableComponent : Component
     public FixedPoint2 WoundableIntegrity;
 
     /// <summary>
-    /// yeah
+    /// Severity thresholds mapping woundable severity levels to their integrity point values.
     /// </summary>
     [DataField(required: true)]
     public Dictionary<WoundableSeverity, FixedPoint2> Thresholds = new();
+
+    /// <summary>
+    /// Pre-sorted version of <see cref="Thresholds"/> in descending order by value.
+    /// Populated on init to avoid per-call OrderByDescending allocations.
+    /// </summary>
+    [Access(typeof(WoundSystem))]
+    public KeyValuePair<WoundableSeverity, FixedPoint2>[]? SortedThresholds;
 
     /// <summary>
     /// How much damage will be healed ACROSS all limb, for example if there are 2 wounds,
