+++ Content.Shared/_Shitmed/Surgery/Wounds/Systems/WoundSystem.cs
@@ -1,13 +1,3 @@
-// SPDX-FileCopyrightText: 2025 Armok <155400926+ARMOKS@users.noreply.github.com>
-// SPDX-FileCopyrightText: 2025 August Eymann <august.eymann@gmail.com>
-// SPDX-FileCopyrightText: 2025 GoobBot <uristmchands@proton.me>
-// SPDX-FileCopyrightText: 2025 Ilya246 <57039557+Ilya246@users.noreply.github.com>
-// SPDX-FileCopyrightText: 2025 Ilya246 <ilyukarno@gmail.com>
-// SPDX-FileCopyrightText: 2025 gluesniffler <159397573+gluesniffler@users.noreply.github.com>
-// SPDX-FileCopyrightText: 2025 gluesniffler <linebarrelerenthusiast@gmail.com>
-//
-// SPDX-License-Identifier: AGPL-3.0-or-later
-
 using Content.Goobstation.Maths.FixedPoint;
 using Content.Shared._Shitmed.Body;
 using Content.Shared._Shitmed.CCVar;
@@ -32,7 +22,7 @@ using Robust.Shared.Network;
 using Robust.Shared.Prototypes;
 using Robust.Shared.Random;
 using Robust.Shared.Timing;
-using System.Linq;
+using Content.Shared.Damage.Prototypes;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -40,6 +30,7 @@ namespace Content.Shared._Shitmed.Medical.Surgery.Wounds.Systems;
 
 public sealed partial class WoundSystem : EntitySystem
 {
+    private Dictionary<string, DamageGroupPrototype?> _damageTypeToGroup = new();
     [Dependency] private readonly IPrototypeManager _prototype = default!;
     [Dependency] private readonly IComponentFactory _factory = default!;
 
@@ -106,6 +97,29 @@ public sealed partial class WoundSystem : EntitySystem
         InitWounding();
         Subs.CVar(_cfg, SurgeryCVars.MedicalHealingTickrate, val => _medicalHealingTickrate = val, true);
         Subs.CVar(_cfg, SurgeryCVars.MinimumTimeBeforeHeal, val => _minimumTimeBeforeHeal = TimeSpan.FromSeconds(val), true);
+
+        BuildDamageTypeToGroupCache();
+        _prototype.PrototypesReloaded += OnPrototypesReloaded;
+    }
+
+    public override void Shutdown()
+    {
+        base.Shutdown();
+        _prototype.PrototypesReloaded -= OnPrototypesReloaded;
+    }
+
+    private void OnPrototypesReloaded(PrototypesReloadedEventArgs args)
+    {
+        if (args.WasModified<DamageGroupPrototype>())
+            BuildDamageTypeToGroupCache();
+    }
+
+    private void BuildDamageTypeToGroupCache()
+    {
+        _damageTypeToGroup.Clear();
+        foreach (var group in _prototype.EnumeratePrototypes<DamageGroupPrototype>())
+            foreach (var damageType in group.DamageTypes)
+                _damageTypeToGroup[damageType] = group;
     }
 
     public override void Update(float frameTime)
@@ -144,15 +158,24 @@ public sealed partial class WoundSystem : EntitySystem
         if (!woundable.Comp.CanHealDamage)
             return;
 
-        var woundsToHeal = GetWoundableWounds(woundable)
-            .Where(wound => CanHealWound(wound, wound))
-            .ToList();
+        var healableCount = 0;
+        foreach (var wound in GetWoundableWounds(woundable))
+        {
+            if (CanHealWound(wound, wound))
+                healableCount++;
+        }
 
-        var healAmount = -woundable.Comp.HealAbility / woundsToHeal.Count;
+        if (healableCount == 0)
+            return;
+
+        var healAmount = -woundable.Comp.HealAbility / healableCount;
         var damageSpecifier = new DamageSpecifier();
         var anythingToHeal = false;
-        foreach (var wound in woundsToHeal)
+        foreach (var wound in GetWoundableWounds(woundable))
         {
+            if (!CanHealWound(wound, wound))
+                continue;
+
             if (wound.Comp.SelfHealMultiplier <= 0)
                 continue;
 
@@ -187,8 +210,6 @@ public sealed partial class WoundSystem : EntitySystem
                     : NetEntity.Invalid,
 
             WoundSeverityPoint = comp.WoundSeverityPoint,
-            WoundableIntegrityMultiplier = comp.WoundableIntegrityMultiplier,
-
             WoundType = comp.WoundType,
 
             DamageGroup = comp.DamageGroup,
@@ -219,7 +240,7 @@ public sealed partial class WoundSystem : EntitySystem
         {
             if (holdingWoundable == EntityUid.Invalid)
             {
-                if (TryComp(holdingWoundable, out WoundableComponent? oldParentWoundable) &&
+                if (TryComp(component.HoldingWoundable, out WoundableComponent? oldParentWoundable) &&
                     TryComp(oldParentWoundable.RootWoundable, out WoundableComponent? oldWoundableRoot))
                 {
                     var ev2 = new WoundRemovedEvent(component, oldParentWoundable, oldWoundableRoot);
@@ -236,13 +257,6 @@ public sealed partial class WoundSystem : EntitySystem
 
                     var ev1 = new WoundAddedEvent(component, parentWoundable, woundableRoot);
                     RaiseLocalEvent(holdingWoundable, ref ev1);
-
-                    var bodyPart = Comp<BodyPartComponent>(holdingWoundable);
-                    if (bodyPart.Body.HasValue)
-                    {
-                        var ev2 = new WoundAddedOnBodyEvent((uid, component), parentWoundable, woundableRoot);
-                        RaiseLocalEvent(bodyPart.Body.Value, ref ev2);
-                    }
                 }
             }
         }
@@ -260,7 +274,6 @@ public sealed partial class WoundSystem : EntitySystem
         }
 
         component.WoundSeverityPoint = state.WoundSeverityPoint;
-        component.WoundableIntegrityMultiplier = state.WoundableIntegrityMultiplier;
 
         if (component.HoldingWoundable != EntityUid.Invalid)
         {
@@ -291,6 +304,18 @@ public sealed partial class WoundSystem : EntitySystem
 
     private void OnWoundableComponentGet(EntityUid uid, WoundableComponent comp, ref ComponentGetState args)
     {
+        var childWoundables = new HashSet<NetEntity>(comp.ChildWoundables.Count);
+        foreach (var woundable in comp.ChildWoundables)
+            childWoundables.Add(TryGetNetEntity(woundable, out var ne) ? ne.Value : NetEntity.Invalid);
+
+        var severityMultipliers = new Dictionary<NetEntity, WoundableSeverityMultiplier>(comp.SeverityMultipliers.Count);
+        foreach (var (key, value) in comp.SeverityMultipliers)
+            severityMultipliers[TryGetNetEntity(key, out var ne) ? ne.Value : NetEntity.Invalid] = value;
+
+        var healingMultipliers = new Dictionary<NetEntity, WoundableHealingMultiplier>(comp.HealingMultipliers.Count);
+        foreach (var (key, value) in comp.HealingMultipliers)
+            healingMultipliers[TryGetNetEntity(key, out var ne) ? ne.Value : NetEntity.Invalid] = value;
+
         var state = new WoundableComponentState
         {
             ParentWoundable = TryGetNetEntity(comp.ParentWoundable, out var parentWoundable) ? parentWoundable : null,
@@ -298,12 +323,7 @@ public sealed partial class WoundSystem : EntitySystem
                 ? rootWoundable.Value
                 : NetEntity.Invalid,
 
-            ChildWoundables =
-                comp.ChildWoundables
-                    .Select(woundable => TryGetNetEntity(woundable, out var ne)
-                        ? ne.Value
-                        : NetEntity.Invalid)
-                    .ToHashSet(),
+            ChildWoundables = childWoundables,
             // Attached and Detached -Woundable events are handled on client with containers
 
             AllowWounds = comp.AllowWounds,
@@ -315,18 +335,8 @@ public sealed partial class WoundSystem : EntitySystem
             WoundableIntegrity = comp.WoundableIntegrity,
             HealAbility = comp.HealAbility,
 
-            SeverityMultipliers =
-                comp.SeverityMultipliers
-                    .Select(multiplier
-                        => (TryGetNetEntity(multiplier.Key, out var ne) ? ne.Value : NetEntity.Invalid,
-                            multiplier.Value))
-                    .ToDictionary(),
-            HealingMultipliers =
-                comp.HealingMultipliers
-                    .Select(multiplier
-                        => (TryGetNetEntity(multiplier.Key, out var ne) ? ne.Value : NetEntity.Invalid,
-                            multiplier.Value))
-                    .ToDictionary(),
+            SeverityMultipliers = severityMultipliers,
+            HealingMultipliers = healingMultipliers,
 
             WoundableSeverity = comp.WoundableSeverity,
         };
@@ -343,10 +353,12 @@ public sealed partial class WoundSystem : EntitySystem
         TryGetEntity(state.RootWoundable, out var rootWoundable);
         component.RootWoundable = rootWoundable ?? EntityUid.Invalid;
 
-        component.ChildWoundables = state.ChildWoundables
-            .Select(x => TryGetEntity(x, out var y) ? y.Value : EntityUid.Invalid)
-            .Where(x => x.Valid)
-            .ToHashSet();
+        component.ChildWoundables.Clear();
+        foreach (var netChild in state.ChildWoundables)
+        {
+            if (TryGetEntity(netChild, out var child) && child.Value.Valid)
+                component.ChildWoundables.Add(child.Value);
+        }
         // Attached and Detached -Woundable events are handled on client with containers
 
         component.AllowWounds = state.AllowWounds;
@@ -357,16 +369,18 @@ public sealed partial class WoundSystem : EntitySystem
         component.HealAbility = state.HealAbility;
         component.Bleeds = state.Bleeds;
 
-        component.SeverityMultipliers =
-            state.SeverityMultipliers
-                .Select(multiplier
-                    => (TryGetEntity(multiplier.Key, out var ne) ? ne.Value : EntityUid.Invalid, multiplier.Value))
-                .ToDictionary();
-        component.HealingMultipliers =
-            state.HealingMultipliers
-                .Select(multiplier
-                    => (TryGetEntity(multiplier.Key, out var ne) ? ne.Value : EntityUid.Invalid, multiplier.Value))
-                .ToDictionary();
+        component.SeverityMultipliers.Clear();
+        foreach (var (key, value) in state.SeverityMultipliers)
+        {
+            if (TryGetEntity(key, out var entityKey))
+                component.SeverityMultipliers[entityKey.Value] = value;
+        }
+        component.HealingMultipliers.Clear();
+        foreach (var (key, value) in state.HealingMultipliers)
+        {
+            if (TryGetEntity(key, out var entityKey))
+                component.HealingMultipliers[entityKey.Value] = value;
+        }
 
         if (component.WoundableIntegrity != state.WoundableIntegrity)
         {
