+++ Content.Shared/_Shitmed/Surgery/Wounds/WoundsSerializable.cs
@@ -1,9 +1,3 @@
-// SPDX-FileCopyrightText: 2025 GoobBot <uristmchands@proton.me>
-// SPDX-FileCopyrightText: 2025 gluesniffler <159397573+gluesniffler@users.noreply.github.com>
-// SPDX-FileCopyrightText: 2025 gluesniffler <linebarrelerenthusiast@gmail.com>
-//
-// SPDX-License-Identifier: AGPL-3.0-or-later
-
 using Content.Shared._Shitmed.Medical.Surgery.Wounds.Components;
 using Content.Goobstation.Maths.FixedPoint;
 using Robust.Shared.Serialization;
@@ -80,9 +74,6 @@ public sealed class WoundVisualizerGroupData : ICloneable
 public record struct WoundAddedEvent(WoundComponent Component, WoundableComponent Woundable, WoundableComponent RootWoundable);
 
 [ByRefEvent]
-public record struct WoundAddedOnBodyEvent(Entity<WoundComponent> Wound, WoundableComponent Woundable, WoundableComponent RootWoundable);
-
-[ByRefEvent]
 public record struct WoundRemovedEvent(WoundComponent Component, WoundableComponent OldWoundable, WoundableComponent OldRootWoundable);
 
 [ByRefEvent]
@@ -93,9 +84,6 @@ public record struct WoundableDetachedEvent(EntityUid ParentWoundableEntity, Wou
 
 [ByRefEvent]
 public record struct WoundSeverityPointChangedEvent(WoundComponent Component, FixedPoint2 OldSeverity, FixedPoint2 NewSeverity, FixedPoint2? Overflow = null);
-
-[ByRefEvent]
-public record struct WoundSeverityPointChangedOnBodyEvent(Entity<WoundComponent> Wound, FixedPoint2 OldSeverity, FixedPoint2 NewSeverity);
 
 [ByRefEvent]
 public record struct WoundSeverityChangedEvent(WoundSeverity OldSeverity, WoundSeverity NewSeverity);
