+++ Resources/Prototypes/Entities/Objects/Tools/bucket.yml
@@ -168,7 +168,7 @@
       - Bucket
   - type: PhysicalComposition #Goobstation - Recycle update
     materialComposition:
-      Plastic: 50
+      Plastic: 10
   - type: DnaSubstanceTrace
   - type: Armor #Goobstation
     coverage:
