+++ Resources/Prototypes/_Shitmed/Entities/Surgery/wounds.yml
@@ -27,7 +27,7 @@
     woundVisibility: Always
     scarWound: BluntScar
     mangleSeverity: Severe
-    integrityMultiplier: 0.5 #heheheha
+
   - type: TraumaInflicter
     severityThreshold: 12
     allowedTraumas:
@@ -65,7 +65,6 @@
     woundType: External
     woundVisibility: Always
     scarWound: PiercingScar
-    integrityMultiplier: 0.1
     mangleSeverity: Severe
   - type: TraumaInflicter
     severityThreshold: 12
@@ -103,7 +102,6 @@
     woundType: External
     woundVisibility: Always
     scarWound: PiercingScar
-    integrityMultiplier: 0.15
     selfHealMultiplier: 0
   - type: TraumaInflicter
     severityThreshold: 6
@@ -138,7 +136,6 @@
     woundType: External
     woundVisibility: Always
     scarWound: SlashScar
-    integrityMultiplier: 0.17
     mangleSeverity: Severe
   - type: TraumaInflicter
     severityThreshold: 9
@@ -172,7 +169,6 @@
     woundType: External
     woundVisibility: Always
     scarWound: BurnScar
-    integrityMultiplier: 0.4 # boiling people
   - type: TraumaInflicter
     severityThreshold: 17
     allowedTraumas:
@@ -206,7 +202,6 @@
     woundType: External
     woundVisibility: Always
     scarWound: BurnScar
-    integrityMultiplier: 0.34 # boiling people
   - type: TraumaInflicter
     severityThreshold: 17
     allowedTraumas:
@@ -237,7 +232,6 @@
     woundType: External
     woundVisibility: Always
     scarWound: BurnScar
-    integrityMultiplier: 0.34
   - type: TraumaInflicter
     severityThreshold: 12
     allowedTraumas:
@@ -260,7 +254,6 @@
     damageType: Holy
     woundType: Internal
     woundVisibility: AdvancedScanner
-    integrityMultiplier: 0
   - type: TraumaInflicter
     severityThreshold: 0
     allowedTraumas:
@@ -283,7 +276,6 @@
     woundType: Internal
     woundVisibility: HandScanner
     scarWound: BurnScar
-    integrityMultiplier: 0.07
   - type: TraumaInflicter
     severityThreshold: 9
     allowedTraumas:
@@ -308,7 +300,6 @@
     woundType: Internal
     woundVisibility: AdvancedScanner
     scarWound: BurnScar
-    integrityMultiplier: 0.15
     selfHealMultiplier: 0
   - type: TraumaInflicter
     severityThreshold: 0
@@ -336,7 +327,6 @@
     woundType: Internal
     woundVisibility: AdvancedScanner
     scarWound: BurnScar
-    integrityMultiplier: 0
   - type: TraumaInflicter
     severityThreshold: 12
     allowedTraumas:
@@ -361,7 +351,6 @@
     woundType: Internal
     woundVisibility: HandScanner
     scarWound: RadiationScar
-    integrityMultiplier: 0
     selfHealMultiplier: 0
   - type: TraumaInflicter
     severityThreshold: 11
@@ -389,7 +378,6 @@
     damageType: Poison
     woundType: Internal
     woundVisibility: HandScanner
-    integrityMultiplier: 0
   - type: TraumaInflicter
     severityThreshold: 7
     allowedTraumas:
